Any/all cleanup and make sure operations can be conditionally reductions
